+++ examples/modl_example.py
@@ -5,7 +5,7 @@ modl = ssbh_data_py.modl_data.read_modl("model.numdlb")
 
 for entry in modl.entries:
     # Each modl entry assigns a material to an object from a mesh file.
-    mesh_object =f'{entry.mesh_object_name}{entry.mesh_object_sub_index}'
+    mesh_object =f'{entry.mesh_object_name}{entry.mesh_object_subindex}'
     print(f'Mesh: {mesh_object}, Material: {entry.material_label}')
 
 # Save any changes made to the modl.